% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savethedate.R
\name{parse_dt}
\alias{parse_dt}
\title{Parse almost anything into a fuzzy datetime}
\usage{
parse_dt(x, ref_date = Sys.Date())
}
\arguments{
\item{x}{A character vector}

\item{ref_date}{A reference date for relative dates (e.g. "yesterday"). Defaults to current date.}
}
\value{
A fuzzy_dt object
}
\description{
Parse almost anything into a fuzzy datetime
}
