% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_the_date.R
\name{parse_dt}
\alias{parse_dt}
\title{Parse almost anything into a fuzzy datetime}
\usage{
parse_dt(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A fuzzy_dt object
}
\description{
Parse almost anything into a fuzzy datetime
}
